src/main/java/com/example/springstarterdemo/discount/RateDiscountPolicy.java
package com.example.springstarterdemo.discount;

import com.example.springstarterdemo.member.Grade;
import com.example.springstarterdemo.member.Member;
import org.springframework.stereotype.Component;

@Component
public class RateDiscountPolicy implements DiscountPolicy{

    private int discountPercent = 10;

    @Override
    public int discount(Member member, int price) {
        if(member.getGrade() == Grade.VIP){
            return price * discountPercent / 100;
        } else {
            return 0;
        }
    }
}
